issue #9241 - cache combobox data
this is the first step in rewriting the xcombobox to solve
problems. future steps include fixing issue #25683 to avoid
data corruption when documents refer to inactive records,
using the model/view system the way qt intended, and perhaps
driving this widget from the source table instead of using
hard-coded queries.
*the caching depends on notify from the database server.*